examples/insurance/code/requirements.txt
google-adk 
litellm 
httpx 
pydantic[dotenv]